requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0
django-ninja==1.1.0
pydantic==2.5.2
pydantic_core==2.14.5
sqlparse==0.4.4
typing_extensions==4.8.0
